dummy-firmware/requirements.txt
pika==1.1.0
retry==0.9.2
paho-mqtt==1.5.0